fix: move all Calculation components to subfolder of DataDimension
