+++ src/components/DataDimension/Calculation/DraggingItem.js
@@ -1,13 +1,13 @@
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { DIMENSION_TYPE_DATA_ELEMENT } from '../../modules/dataTypes.js'
-import { getIcon } from '../../modules/dimensionListItem.js'
+import { DIMENSION_TYPE_DATA_ELEMENT } from '../../../modules/dataTypes.js'
+import { getIcon } from '../../../modules/dimensionListItem.js'
 import {
     EXPRESSION_TYPE_DATA,
     EXPRESSION_TYPE_NUMBER,
     EXPRESSION_TYPE_OPERATOR,
-} from '../../modules/expressions.js'
+} from '../../../modules/expressions.js'
 import styles from './styles/DraggingItem.style.js'
 import formulaItemStyles from './styles/FormulaItem.style.js'
