+++ src/components/DataDimension/Calculation/FormulaField.js
@@ -4,7 +4,7 @@ import { useDroppable } from '@dnd-kit/core'
 import { SortableContext } from '@dnd-kit/sortable'
 import PropTypes from 'prop-types'
 import React from 'react'
-import FormulaIcon from '../../assets/FormulaIcon.js'
+import FormulaIcon from '../../../assets/FormulaIcon.js'
 import DropZone from './DropZone.js'
 import FormulaItem from './FormulaItem.js'
 import styles from './styles/FormulaField.style.js'
