+++ src/components/DataDimension/Calculation/styles/DataElementSelector.style.js
