+++ src/components/DataDimension/Calculation/styles/DropZone.style.js
