+++ src/components/DataDimension/Calculation/styles/MathOperatorSelector.style.js
