+++ src/components/DataDimension/Calculation/styles/Operator.style.js
